Moved reporters to a separate package
Part of #588
