+++ src/main/resources/com/sap/oss/phosphor/fosstars/tool/report/MarkdownProjectDetailsTemplate.md
