+++ src/main/resources/com/sap/oss/phosphor/fosstars/tool/report/OssSecurityRatingMarkdownReporterMainTemplate.md
